% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gb_file.R
\name{FasExtract}
\alias{FasExtract}
\title{Genome extracter}
\usage{
FasExtract(gb)
}
\arguments{
\item{gb}{a large character vector returned by function
\code{\link{fetch.gb}} or \code{\link{read.gb}}}
}
\value{
The genome of the sequence that is deposited at the end of the
GeneBank file in fasta format
}
\description{
Extracting the fasta format chloroplast genome from the GeneBank File
}
