% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gb_file.R
\name{fetch.gb}
\alias{fetch.gb}
\title{Fetch GB files online}
\usage{
fetch.gb(GI, read = TRUE)
}
\arguments{
\item{GI}{GenInfo Identifier(GI) for interested genome file.}

\item{read}{A logical value. If it is \code{TRUE}, the function will read
the fetched GB file into current R session environment.}
}
\value{
If \code{read} is set as \code{TRUE}. The function return a large
character vector which contains the content of GB file for input \code{GI}
from GeneBank.
}
\description{
Fetching GB file from GenBank database by GenInfo Identifier (GI) or
accession number and writeing it to a file named "GI.gb" in current work
directory.
}
