% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_IR_info.R
\name{fun}
\alias{fun}
\title{Auxiliary function for parallely detecting phase difference}
\usage{
fun(shifter, genome)
}
\arguments{
\item{shifter}{a integer. It indicates the number of nucleotides which will
be shifted to the end of the genome sequence.}

\item{genome}{a large character vector. It}
}
\value{

}
\description{
the function to be passed to the slaves for the parallel computing, the
output is with max 10 second either NA or the phase.detector
}
