% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_info.R
\name{geneTable}
\alias{geneTable}
\title{Generate gene table}
\usage{
geneTable(gb, genome)
}
\arguments{
\item{gb}{parsed gbfile}

\item{genome}{a DNAstring object. It contains the genome sequence.}
}
\value{
a data frame.
}
\description{
Generate gene table
}
