% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_info.R
\name{geneTableParsed}
\alias{geneTableParsed}
\title{Generate gene table from parsed gb file}
\usage{
geneTableParsed(gb, genome)
}
\arguments{
\item{gb}{A list containing parsed GB file information. It is generated by
function \code{\link[genbankr]{parseGenBank}}}

\item{genome}{a DNAstring object. It contains the genome sequence.}
}
\value{
a data frame. It contains information of genes.
}
\description{
Generate gene table from parsed gb file
}
