% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_info.R
\name{geneTableRead}
\alias{geneTableRead}
\title{Generate gene table from GenBankRecord}
\usage{
geneTableRead(gb, genome)
}
\arguments{
\item{gb}{Formalclass GenBankRecord. It is generated by function
\code{\link[genbankr]{readGenBank}}}

\item{genome}{a DNAstring object. It contains the genome sequence.}
}
\value{
a data frame. It contains information of genes.
}
\description{
Generate gene table from GenBankRecord
}
