% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_IR_info.R
\name{IRinfo}
\alias{IRinfo}
\title{IR information}
\usage{
IRinfo(genome, nCPU = 4)
}
\arguments{
\item{genome}{The plastid genome of a species as a simple vector of nucleotides.}

\item{nCPU}{number of CPU used for detecting phase difference of the two
inverted genomes.}
}
\value{
a vector of four elements as the start of the first and second IR region,
their length and the total lenght of the genome, respectively
}
\description{
Detecting the starts and the length of the IR regions on the chloroplast genome
}
