% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_IR_info.R
\name{IRtab}
\alias{IRtab}
\title{Make the table of the IR information}
\usage{
IRtab(GBFiles)
}
\arguments{
\item{GBFiles}{A vector of characters. It contains the GIs or GenBank
accessions for the interested species' chloroplast genome files.}
}
\value{
A matrix with the IR information for each input species
}
\description{
This function generates a table for information about IR region from multiple
GenBank files.
}
