% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_IR_info.R
\name{phase.detector}
\alias{phase.detector}
\title{Detect the phase difference of the two inverted genomes}
\usage{
phase.detector(genome, nCPU = 1)
}
\arguments{
\item{genome}{A large character vector. It is the plastid genome of a species
as a simple vector of nucleotides.}

\item{nCPU}{a integer. It represent the number of CPU will be used to
run the function.}
}
\value{
A integer. It indicates the start position of the first IR region.
}
\description{
Detect the phase difference of the two inverted genomes
}
