% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{PlotGenome}
\alias{PlotGenome}
\title{Generate Genome plot}
\usage{
PlotGenome(
  plot_tables,
  save = TRUE,
  file.type = "pdf",
  file.name = NULL,
  shadow = TRUE,
  legend = TRUE,
  ssc.converse = FALSE,
  background = "grey90",
  gc.color = "grey30",
  gc.background = "grey70",
  info.background = "black",
  ir.color = "#2F3941",
  shadow.color = "#0000FF20",
  ssc.color = "#82B6E2",
  lsc.color = "#299E96",
  height = c(0.1, 0.2, 0.07)
)
}
\arguments{
\item{plot_tables}{a list contains information of IR region, gene, and gc
count of the genome. It can be generated by function \code{PlotTab}.}

\item{save}{A logical value. If it is \code{TRUE}, the plot will be saved in
work directory.}

\item{file.type}{A charactor. It indicates the format of the file in which
the plot will be saved. Options are: "pdf", "png", "jpeg","bmp", "tiff".
By default, it is set as "pdf".}

\item{file.name}{A charactor. It indicates the name of the file in which
the plot will be saved. By default, it is set as the specie's name.}

\item{shadow}{A logical value. If it is \code{TRUE}, the section of IR
regions will be highlighted by shadows.}

\item{legend}{A logical value. If it is \code{TRUE}, the legend for gene
colors will be shown.}

\item{ssc.converse}{A logical value. If it is \code{TRUE}, the SSC region
will be converted to its reverse complementary version}

\item{background}{A character. It indicates the color for the background of
entire plot area.}

\item{gc.color}{A character. It indicates the color for the lines in gc count
plot.}

\item{gc.background}{A character. It indicates the color for the background
of gc count plot area.}

\item{info.background}{A character. It indicates the color for the background
of central area where species' information was shown.}

\item{ir.color}{A character. It indicates the color for the background
of IR sectors.}

\item{shadow.color}{A character. It indicates the color for the shadow
casted from IR sectors.}

\item{ssc.color}{A character. It indicates the color for the background
of SSC sectors.}

\item{lsc.color}{A character. It indicates the color for the background
of LSC sectors.}

\item{height}{A vector of numeric value. The elements of it indicat the
height of gene layer, GC count layer and IR region layer, respectively.
Default setting is "0.1, 0.2, 0.07". The taltal circle plot region always
has a radius of 1, so a height of 0.1 means 10% of the circle radius.}
}
\value{
A plot for chloroplast genome.
}
\description{
Generate Genome plot
}
