% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{PlotTab}
\alias{PlotTab}
\title{Generate table for PlotGenome function}
\usage{
PlotTab(gbfile, local.file = FALSE, gc.window = 100)
}
\arguments{
\item{gbfile}{A character. It is the path to a local GB file or GenBank
accession for the interested specie's chloroplast genome. If you want to use
GenBank accession here, please set parameter \code{local.file} as \code{FALSE}.}

\item{local.file}{A logical value. If it is true the function will try to
read a gb file from local directory. If it is true it will regard the value
passed to parameter \code{gbfile} as GenBank accession and try ro fetch
corresponding data from GeneBank.}

\item{gc.window}{A integer. It indicates the window sized for plot
GC count.}
}
\value{
A list. It contains 3 tables:
\itemize{
  \item ir_table: a data frame contains information for IR region.
  \item gene_table: a data frame contains information of genes.
  \item gc_count: a data frame contains GC count information.
}
}
\description{
Generate table for PlotGenome function
}
