% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gb_file.R
\name{rdnFixer}
\alias{rdnFixer}
\title{Randomly fill nucleotides}
\usage{
rdnFixer(genome)
}
\arguments{
\item{genome}{A DNAString object. It is a class defined in \code{Biostrings}
package. It contains the whole genome sequence of a specie.}
}
\value{
A large character vector. It countains the nucleotide sequence with
all letters except "a", "t", "c", "g" changed.
}
\description{
This fuction change liters "u", "r", "y", "s", "w", "k", "m", "b", "d", "h",
"v", "n" or "-" in genome sequence (randomly) into corresponding nucleotide
"a", "t", "c" or "g".
}
