% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gb_file.R
\name{read.gb}
\alias{read.gb}
\title{Read the local GB file}
\usage{
read.gb(file)
}
\arguments{
\item{file}{The path to the GB file which is in GeneBank format}
}
\value{
a large character vector returned by function
\code{\link{fetch.gb}} or \code{\link{read.gb}}
}
\description{
Reading the GB file in the working directory producing the RGB format file
needed for the functions of the package
}
