% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gb_file.R
\name{sp.name}
\alias{sp.name}
\title{Extract specie's name from GB file}
\usage{
sp.name(definition, text = FALSE)
}
\arguments{
\item{gb}{a large character vector returned by function
\code{\link{fetch.gb}} or \code{\link{read.gb}}}
}
\value{
a character contains the specie's name.
}
\description{
Extract specie's name from GB file
}
