% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{PlotGenome}
\alias{PlotGenome}
\title{Generate Genome plot}
\usage{
PlotGenome(
  gbfile,
  save = TRUE,
  file.type = "pdf",
  file.name = NULL,
  gc.window = 100,
  ssc.converse = FALSE,
  shadow = TRUE
)
}
\arguments{
\item{gbfile}{A character. It is the GI or GenBank accession for the
interested specie's chloroplast genome files.}

\item{save}{A logical value. If it is \code{TRUE}, the plot will be saved in
work directory.}

\item{file.type}{A charactor. It indicates the format of the file in which
the plot will be saved. Options are: "pdf", "png", "jpeg","bmp", "tiff".
By default, it is set as "pdf".}

\item{file.name}{A charactor. It indicates the name of the file in which
the plot will be saved. By default, it is set as the specie's name.}

\item{gc.window}{A integer. It indicates the window sized for plot
GC count.}
}
\value{
A plot for chloroplast genome.
}
\description{
Generate Genome plot
}
