% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{PlotGenome}
\alias{PlotGenome}
\title{Generate Genome plot}
\usage{
PlotGenome(
  plot_tables,
  save = TRUE,
  file.type = "pdf",
  file.name = NULL,
  shadow = TRUE,
  shadow.color = "#0000FF20",
  legend = TRUE,
  ssc.converse = FALSE,
  background = "grey90",
  height = c(0.1, 0.2, 0.07),
  gc.color = "grey30",
  gc.background = "grey70",
  info.background = "black",
  ir.color = "#2F3941",
  ssc.color = "#82B6E2",
  lsc.color = "#299E96",
  ir_gc = TRUE,
  gc_per_gene = TRUE,
  pseudo = TRUE,
  cu_bias = TRUE,
  customize_ring1 = NULL,
  customize_ring2 = NULL
)
}
\arguments{
\item{plot_tables}{a list contains information of IR region, gene, and gc
count of the genome. It can be generated by function \code{\link{PlotTab}}.}

\item{save}{A logical value. If it is \code{TRUE}, the plot will be saved in
work directory.}

\item{file.type}{A charactor. It indicates the format of the file in which
the plot will be saved. Options are: "pdf", "png", "jpeg","bmp", "tiff".
By default, it is set as "pdf".}

\item{file.name}{A charactor. It indicates the name of the file in which
the plot will be saved. By default, it is set as the specie's name.}

\item{shadow}{A logical value. If it is \code{TRUE}, the section of IR
regions will be highlighted by shadows.}

\item{shadow.color}{A character. It indicates the color for the shadow
casted from IR sectors.}

\item{legend}{A logical value. If it is \code{TRUE}, the legend for gene
colors will be shown.}

\item{ssc.converse}{A logical value. If it is \code{TRUE}, the SSC region
will be converted to its reverse complementary version}

\item{background}{A character. It indicates the color for the background of
entire plot area.}

\item{height}{A vector of numeric value. The elements of it indicat the
height of gene layer, GC count layer and IR region layer, respectively.
Default setting is "0.1, 0.2, 0.07". The taltal circle plot region always
has a radius of 1, so a height of 0.1 means 10\% of the circle radius.}

\item{gc.color}{A character. It indicates the color for the lines in gc count
plot.}

\item{gc.background}{A character. It indicates the color for the background
of gc count plot area.}

\item{info.background}{A character. It indicates the color for the background
of central area where species' information was shown.}

\item{ir.color}{A character. It indicates the color for the background
of IR sectors.}

\item{ssc.color}{A character. It indicates the color for the background
of SSC sectors.}

\item{lsc.color}{A character. It indicates the color for the background
of LSC sectors.}

\item{ir_gc}{A logical value. If it is \code{TRUE}, the gc content for
IR, LSC, and SSC regions will be plot with deep colors.}

\item{gc_per_gene}{A logical value. If it is \code{TRUE}, the gc content for
each gene will be plot on the gene layer with deep colors.}

\item{pseudo}{A logical value. If it is \code{TRUE}, the pseudo genes (if
there is some in the species) will be marked with a "*" at the end of the
labels.}

\item{cu_bias}{A logical value. If it is \code{TRUE}, the condon usage bias
metric for each gene will be shown in the labels. The metric “Measure
Independent of Length and Composition (MILC)” was used for evaluate the bias.
You can find more details in the reference paper.}

\item{customize_ring1}{coming soon}

\item{customize_ring2}{coming soon}
}
\value{
A plot for chloroplast genome.
}
\description{
Generate Genome plot
}
\references{
Supek, Fran, and Kristian Vlahovicek. “Comparison of codon usage
measures and their applicability in prediction of microbial gene
expressivity.” BMC bioinformatics vol. 6 182. 19 Jul.
2005, doi:10.1186/1471-2105-6-182
}
