% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{PlotTab}
\alias{PlotTab}
\title{Generate table for PlotGenome function}
\usage{
PlotTab(gbfile, local.file = FALSE, gc.window = 100)
}
\arguments{
\item{gbfile}{A character. It is the GI or GenBank accession for the
interested specie's chloroplast genome files.}

\item{local.file}{A logical value. If it is \code{TRUE}, the value passed to
\code{gbfile} should be a local GB file.}

\item{gc.window}{A integer. It indicates the window sized for plot
GC count.}
}
\value{
A list. It contains 3 tables:
\itemize{
  \item ir_table: a data frame contains information for IR region.
  \item gene_table: a data frame contains information of genes.
  \item gc_count: a data frame contains GC count information.
}
}
\description{
Generate table for PlotGenome function
}
