% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gb_file.R
\name{sp.name}
\alias{sp.name}
\title{Extract specie's name from GB file}
\usage{
sp.name(definition)
}
\arguments{
\item{definition}{a charactor. It contain the definition field from the
GenBank file.}
}
\value{
a character contains the specie's name.
}
\description{
Extract specie's name from GB file
}
